Restructure to subclass PawnStorageNutrition
