+++ 1.5/Defs/Buildings/Buildings_Structure.xml
@@ -163,7 +163,7 @@
                 <lightEffect>true</lightEffect>
                 <storageStation>PS_BatteryFarm</storageStation>
             </li>
-            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmNutrition">
+            <li Class="PawnStorages.CompProperties_PawnStorageNutrition">
                 <HasTip>false</HasTip>
                 <MainTex>Things/Building/Production/PS_BatteryFarm</MainTex>
                 <AltTex>Things/Building/Production/PS_BatteryFarmRustic</AltTex>
@@ -268,7 +268,7 @@
                 <lightEffect>true</lightEffect>
                 <storageStation>PS_BatteryFarm</storageStation>
             </li>
-            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmNutrition">
+            <li Class="PawnStorages.CompProperties_PawnStorageNutrition">
                 <HasTip>true</HasTip>
                 <TipScale>5</TipScale>
                 <MainTex>Things/Building/Production/PS_BreedingDomes</MainTex>
