+++ 1.5/Defs/Buildings/Buildings_Structure_Rustic.xml
@@ -48,7 +48,7 @@
                 <lightEffect>true</lightEffect>
                 <storageStation>PS_BatteryFarm</storageStation>
             </li>
-            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmNutrition">
+            <li Class="PawnStorages.CompProperties_PawnStorageNutrition">
                 <HasTip>false</HasTip>
             </li>
             <li Class="PawnStorages.Farm.Comps.CompProperties_FarmProducer">
@@ -151,7 +151,7 @@
                 <lightEffect>true</lightEffect>
                 <storageStation>PS_BatteryFarm</storageStation>
             </li>
-            <li Class="PawnStorages.Farm.Comps.CompProperties_FarmNutrition">
+            <li Class="PawnStorages.CompProperties_PawnStorageNutrition">
                 <HasTip>true</HasTip>
                 <TipScale>5</TipScale>
                 <MainTex>Things/Building/Production/PS_BreedingDomesRustic</MainTex>
