+++ 1.5/Source/PawnStorages/PawnStoragesModule-VEF/CompProperties_PawnStorageNutrition.cs
@@ -1,0 +1,9 @@
+﻿namespace PawnStorages.VEF;
+
+public class CompProperties_PipedPawnStorageNutrition : CompProperties_PawnStorageNutrition
+{
+    public CompProperties_PipedPawnStorageNutrition()
+    {
+        compClass = typeof(CompPipedPawnStorageNutrition);
+    }
+}
