+++ modList.xml
@@ -18,6 +18,7 @@
         <li>vanillaexpanded.vchemfuele</li>
         <li>vanillaexpanded.helixiengas</li>
         <li>bambaryla.vcepu</li>
+        <li>smashphil.xmlpatchhelper</li>
     </activeMods>
     <knownExpansions>
         <li>ludeon.rimworld</li>
